--- transfer2/config.yaml
@@ -1,20 +1,0 @@
-# config.yaml
-
-# Host PC에서 Guest VM1(우분투)로 전송 설정
-host_to_vm1:
-  host: ""         # VM1의 IP
-  port: 22
-  username: ""
-  password: ""
-  remote_path: ""  # VM1에서 파일이 저장될 경로
-
-# VM1에서 Guest VM2(윈도우)로 전송 설정
-vm1_to_vm2:
-  host: ""         # VM2의 IP
-  port: 22
-  username: ""          # 예: Windows의 경우 "winuser"
-  password: ""
-  remote_path: ""   # VM2에서 파일이 저장될 경로
-
-# VM1에서 임시 파일(스크립트, config 파일 등)을 저장할 디렉토리
-vm1_temp_dir: "/home/vm1user/temporary/"  # 이 디렉토리가 없으면 자동으로 생성합니다.
