--- transfer_test/config.yaml
@@ -1,34 +1,0 @@
-# config.yaml
-
-# 전송할 파일 기본 경로 (옵션)
-default_file: 'D:\goorm\FinalProject\goorm_final_pjt\transfer_test\analysis_test.json'
-
-# Host PC에서 Guest VM1(우분투)로 전송 설정
-host_to_vm1:
-  host: 222.114.170.35
-  port: 22
-  username: pyh
-  password: '1234'
-  remote_path: "/home/pyh/Desktop/"  # 파일은 이 디렉토리에 저장됨
-
-# Guest VM1(우분투)에서 Guest VM2(윈도우)로 전송 설정
-vm1_to_vm2:
-  host: "192.168.1.102"         # Guest VM2의 IP
-  port: 22
-  username: "vm2_user"          # 예: Windows의 경우 "winuser"
-  password: "vm2_password"
-  remote_path: "C:/analysis/"   # Windows에서는 이 경로에 저장됨
-
-# Guest VM2(윈도우)에서 분석 보고서(JSON)를 Guest VM1(우분투)로 전송 설정
-vm2_to_vm1:
-  host: 222.114.170.35
-  port: 22
-  username: pyh
-  password: '1234'
-  remote_path: "/home/pyh/Desktop/"
-
-# AWS ELK 전송 설정
-elk:
-  endpoint: "http://3.36.50.236:9200/malware-analysis-2025.02.02/_doc"
-  headers:
-    Content-Type: "application/json"
